+++ packages/koop-core/test/model-test.js
@@ -30,7 +30,7 @@ describe('Tests for models/index', function () {
 
       retrieveSpy.should.be.calledOnce()
       retrieveSpy.firstCall.args.length.should.equal(3)
-      retrieveSpy.firstCall.args[0].should.equal('test-provider')
+      retrieveSpy.firstCall.args[0].should.equal('test-provider::data')
       retrieveSpy.firstCall.args[1].should.be.an.Object().and.be.empty()
       retrieveSpy.firstCall.args[2].should.be.an.Function()
       pullSpy.should.be.calledOnce()
@@ -55,7 +55,7 @@ describe('Tests for models/index', function () {
 
       retrieveSpy.should.be.calledOnce()
       retrieveSpy.firstCall.args.length.should.equal(3)
-      retrieveSpy.firstCall.args[0].should.equal('test-provider::host-param::id-param::layer-param')
+      retrieveSpy.firstCall.args[0].should.equal('test-provider::host-param::id-param::layer-param::data')
       retrieveSpy.firstCall.args[1].should.be.an.Object().and.be.empty()
       retrieveSpy.firstCall.args[2].should.be.an.Function()
       pullSpy.should.be.calledOnce()
@@ -80,7 +80,7 @@ describe('Tests for models/index', function () {
       model.pull({ url: 'domain/test-provider', query: {} }, pullSpy)
       retrieveSpy.should.be.calledOnce()
       retrieveSpy.firstCall.args.length.should.equal(3)
-      retrieveSpy.firstCall.args[0].should.equal('custom-key')
+      retrieveSpy.firstCall.args[0].should.equal('custom-key::data')
       retrieveSpy.firstCall.args[1].should.be.an.Object().and.be.empty()
       retrieveSpy.firstCall.args[2].should.be.an.Function()
       pullSpy.should.be.calledOnce()
@@ -161,6 +161,182 @@ describe('Tests for models/index', function () {
       pullCallbackSpy.firstCall.args.length.should.equal(2)
       should.not.exist(pullCallbackSpy.firstCall.args[0])
       pullCallbackSpy.firstCall.args[1].should.be.an.Object().and.deepEqual({ metadata: { food: 'bar' } })
+    })
+  })
+
+  describe('fetch layer metadata', function () {
+    afterEach(function () {
+      // reset the getLayer() function to default
+      Model.prototype.getLayer = undefined
+    })
+
+    it('should throw an error if the getLayer() function is not implemented', function () {
+      const retrieveSpy = sinon.spy(function (key, queryParams, callback) {
+        callback(new Error('not found'))
+      })
+      const callbackSpy = sinon.spy()
+
+      // create a model with mocked cache "retrieve" function
+      const model = new Model({
+        cache: {
+          retrieve: retrieveSpy
+        }
+      })
+
+      model.pullLayer({ url: 'domain/test-provider', params: {}, query: {} }, callbackSpy)
+
+      retrieveSpy.should.be.calledOnce()
+      callbackSpy.should.be.calledOnce()
+      callbackSpy.firstCall.args.length.should.equal(1)
+      callbackSpy.firstCall.args[0].should.be.an.Error()
+    })
+
+    it('should try to fetch from cache', function () {
+      const retrieveSpy = sinon.spy(function (key, queryParams, callback) {
+        callback(null, {})
+      })
+      const getLayerSpy = sinon.spy(function (req, callback) {
+        callback(null, {})
+      })
+      const callbackSpy = sinon.spy()
+
+      Model.prototype.getLayer = getLayerSpy
+
+      // create a model with mocked cache "retrieve" function
+      const model = new Model({
+        cache: {
+          retrieve: retrieveSpy
+        }
+      })
+
+      model.pullLayer({ url: 'domain/test-provider', params: {}, query: {} }, callbackSpy)
+
+      retrieveSpy.should.be.calledOnce()
+      retrieveSpy.firstCall.args.length.should.equal(3)
+      retrieveSpy.firstCall.args[0].should.equal('test-provider::layer')
+      retrieveSpy.firstCall.args[1].should.be.an.Object().and.be.empty()
+      retrieveSpy.firstCall.args[2].should.be.an.Function()
+      getLayerSpy.should.not.be.called()
+      callbackSpy.should.be.calledOnce()
+      callbackSpy.firstCall.args.length.should.equal(2)
+      should.not.exist(callbackSpy.firstCall.args[0])
+      callbackSpy.firstCall.args[1].should.be.an.Object()
+    })
+
+    it('should call the getLayer() function is cache misses', function () {
+      const retrieveSpy = sinon.spy(function (key, queryParams, callback) {
+        callback(new Error('not found'))
+      })
+      const getLayerSpy = sinon.spy(function (req, callback) {
+        callback(null, {})
+      })
+      const callbackSpy = sinon.spy()
+
+      Model.prototype.getLayer = getLayerSpy
+
+      // create a model with mocked cache "retrieve" function
+      const model = new Model({
+        cache: {
+          retrieve: retrieveSpy
+        }
+      })
+
+      model.pullLayer({ url: 'domain/test-provider', params: {}, query: {} }, callbackSpy)
+
+      retrieveSpy.should.be.calledOnce()
+      getLayerSpy.should.be.calledOnce()
+      callbackSpy.should.be.calledOnce()
+      callbackSpy.firstCall.args.length.should.equal(2)
+      should.not.exist(callbackSpy.firstCall.args[0])
+      callbackSpy.firstCall.args[1].should.be.an.Object()
+    })
+  })
+
+  describe('fetch catalog metadata', function () {
+    afterEach(function () {
+      // reset the getCatalog() function to default
+      Model.prototype.getCatalog = undefined
+    })
+
+    it('should throw an error if the getCatalog() function is not implemented', function () {
+      const retrieveSpy = sinon.spy(function (key, queryParams, callback) {
+        callback(new Error('not found'))
+      })
+      const callbackSpy = sinon.spy()
+
+      // create a model with mocked cache "retrieve" function
+      const model = new Model({
+        cache: {
+          retrieve: retrieveSpy
+        }
+      })
+
+      model.pullCatalog({ url: 'domain/test-provider', params: {}, query: {} }, callbackSpy)
+
+      retrieveSpy.should.be.calledOnce()
+      callbackSpy.should.be.calledOnce()
+      callbackSpy.firstCall.args.length.should.equal(1)
+      callbackSpy.firstCall.args[0].should.be.an.Error()
+    })
+
+    it('should try to fetch from cache', function () {
+      const retrieveSpy = sinon.spy(function (key, queryParams, callback) {
+        callback(null, {})
+      })
+      const getCatalogSpy = sinon.spy(function (req, callback) {
+        callback(null, {})
+      })
+      const callbackSpy = sinon.spy()
+
+      Model.prototype.getCatalog = getCatalogSpy
+
+      // create a model with mocked cache "retrieve" function
+      const model = new Model({
+        cache: {
+          retrieve: retrieveSpy
+        }
+      })
+
+      model.pullCatalog({ url: 'domain/test-provider', params: {}, query: {} }, callbackSpy)
+
+      retrieveSpy.should.be.calledOnce()
+      retrieveSpy.firstCall.args.length.should.equal(3)
+      retrieveSpy.firstCall.args[0].should.equal('test-provider::catalog')
+      retrieveSpy.firstCall.args[1].should.be.an.Object().and.be.empty()
+      retrieveSpy.firstCall.args[2].should.be.an.Function()
+      getCatalogSpy.should.not.be.called()
+      callbackSpy.should.be.calledOnce()
+      callbackSpy.firstCall.args.length.should.equal(2)
+      should.not.exist(callbackSpy.firstCall.args[0])
+      callbackSpy.firstCall.args[1].should.be.an.Object()
+    })
+
+    it('should call the getCatalog() function is cache misses', function () {
+      const retrieveSpy = sinon.spy(function (key, queryParams, callback) {
+        callback(new Error('not found'))
+      })
+      const getCatalogSpy = sinon.spy(function (req, callback) {
+        callback(null, {})
+      })
+      const callbackSpy = sinon.spy()
+
+      Model.prototype.getCatalog = getCatalogSpy
+
+      // create a model with mocked cache "retrieve" function
+      const model = new Model({
+        cache: {
+          retrieve: retrieveSpy
+        }
+      })
+
+      model.pullCatalog({ url: 'domain/test-provider', params: {}, query: {} }, callbackSpy)
+
+      retrieveSpy.should.be.calledOnce()
+      getCatalogSpy.should.be.calledOnce()
+      callbackSpy.should.be.calledOnce()
+      callbackSpy.firstCall.args.length.should.equal(2)
+      should.not.exist(callbackSpy.firstCall.args[0])
+      callbackSpy.firstCall.args[1].should.be.an.Object()
     })
   })
 })